Fix attacher issues
