+++ packages/dev/core/src/AudioV2/abstractAudio/components/spatialAudioAttacherComponent.ts
@@ -1,25 +1,15 @@
 import type { Camera } from "../../../Cameras/camera";
 import type { AbstractMesh } from "../../../Meshes/abstractMesh";
 import type { TransformNode } from "../../../Meshes/transformNode";
+import type { Node } from "../../../node";
 import type { Nullable } from "../../../types";
 import type { _AbstractSpatialAudioAttacher } from "../spatialAttachers/abstractSpatialAudioAttacher";
+import { _SpatialAudioAttachedEntity, SpatialAudioAttachmentType } from "../spatialAttachers/abstractSpatialAudioAttacher";
 import { _SpatialAudioCameraAttacher } from "../spatialAttachers/spatialAudioCameraAttacher";
 import { _SpatialAudioMeshAttacher } from "../spatialAttachers/spatialAudioMeshAttacher";
 import { _SpatialAudioTransformNodeAttacher } from "../spatialAttachers/spatialAudioTransformNodeAttacher";
 import type { _SpatialAudioSubNode } from "../subNodes/spatialAudioSubNode";
 import type { _SpatialAudioListener } from "../subProperties/spatialAudioListener";
-
-export const _SpatialAudioAttachedEntity = {
-    CAMERA: "Camera",
-    MESH: "Mesh",
-    TRANSFORM_NODE: "TransformNode",
-} as const;
-
-export const enum SpatialAudioAttachmentType {
-    POSITION = 1,
-    ROTATION = 2,
-    POSITION_AND_ROTATION = 3,
-}
 
 /**
  * Provides a common interface for attaching an audio listener or source to a specific entity, ensuring only one entity
@@ -27,8 +17,10 @@ export const enum SpatialAudioAttachmentType {
  * @internal
  */
 export class _SpatialAudioAttacherComponent {
-    private _attachedEntity: Nullable<AbstractMesh | Camera | TransformNode> = null;
     private _attacher: Nullable<_AbstractSpatialAudioAttacher> = null;
+
+    /** @internal */
+    public attachedNode: Nullable<Node> = null;
 
     /**
      * The type of attachment to use; position, rotation, or both.
@@ -51,7 +43,7 @@ export class _SpatialAudioAttacherComponent {
      * attached to a scene.
      */
     public get attachedCamera(): Nullable<Camera> {
-        return this._attacher?.getClassName() === _SpatialAudioAttachedEntity.CAMERA ? (this._attachedEntity as Camera) : null;
+        return this._attacher?.getAttacherType() === _SpatialAudioAttachedEntity.CAMERA ? (this.attachedNode as Camera) : null;
     }
 
     public set attachedCamera(value: Nullable<Camera>) {
@@ -67,7 +59,7 @@ export class _SpatialAudioAttacherComponent {
      * attached to a mesh.
      */
     public get attachedMesh(): Nullable<AbstractMesh> {
-        return this._attacher?.getClassName() === _SpatialAudioAttachedEntity.MESH ? (this._attachedEntity as AbstractMesh) : null;
+        return this._attacher?.getAttacherType() === _SpatialAudioAttachedEntity.MESH ? (this.attachedNode as AbstractMesh) : null;
     }
 
     public set attachedMesh(value: Nullable<AbstractMesh>) {
@@ -83,7 +75,7 @@ export class _SpatialAudioAttacherComponent {
      * not attached to a transform node.
      */
     public get attachedTransformNode(): Nullable<TransformNode> {
-        return this._attacher?.getClassName() === _SpatialAudioAttachedEntity.TRANSFORM_NODE ? (this._attachedEntity as TransformNode) : null;
+        return this._attacher?.getAttacherType() === _SpatialAudioAttachedEntity.TRANSFORM_NODE ? (this.attachedNode as TransformNode) : null;
     }
 
     public set attachedTransformNode(value: Nullable<TransformNode>) {
@@ -140,7 +132,7 @@ export class _SpatialAudioAttacherComponent {
     private _resetAttachedEntity(entity: Nullable<AbstractMesh | Camera | TransformNode>, attacherClassName: string): void {
         this.detach();
 
-        this._attachedEntity = entity;
+        this.attachedNode = entity;
 
         if (!entity) {
             this._attacher = null;
