+++ packages/dev/core/src/AudioV2/abstractAudio/spatialAttachers/spatialAudioCameraAttacher.ts
@@ -1,18 +1,11 @@
 import type { Camera } from "../../../Cameras/camera";
 import type { Vector3 } from "../../../Maths/math.vector";
 import type { Nullable } from "../../../types";
-import type { _SpatialAudioAttacherComponent } from "../components/spatialAudioAttacherComponent";
-import { _AbstractSpatialAudioAttacher } from "./abstractSpatialAudioAttacher";
+import { _AbstractSpatialAudioAttacher, _SpatialAudioAttachedEntity } from "./abstractSpatialAudioAttacher";
 
 /** @internal */
 export class _SpatialAudioCameraAttacher extends _AbstractSpatialAudioAttacher {
-    protected override _node: Nullable<Camera> = null;
-
-    /** @internal */
-    public constructor(attacherComponent: _SpatialAudioAttacherComponent) {
-        super(attacherComponent);
-        this.node = attacherComponent.attachedCamera;
-    }
+    protected override _node: Nullable<Camera>;
 
     protected get _attachedPosition(): Vector3 {
         if (this._node) {
@@ -24,7 +17,7 @@ export class _SpatialAudioCameraAttacher extends _AbstractSpatialAudioAttacher {
     }
 
     /** @internal */
-    public getClassName(): string {
-        return "_SpatialAudioCameraAttacher";
+    public getAttacherType(): string {
+        return _SpatialAudioAttachedEntity.CAMERA;
     }
 }
